docs : users 관련 VO 수정

+++ src/main/java/com/coffice/app/users/AnnualLeaveVO.java
@@ -1,6 +1,6 @@
 package com.coffice.app.users;
 
-import java.math.BigDecimal;
+
 import java.sql.Date;
 
 import lombok.Getter;
@@ -14,8 +14,8 @@ public class AnnualLeaveVO {
 	
 	private Integer leaveId;
 	private Integer leaveYear;
-	private BigDecimal totalLeave;
-	private BigDecimal usedLeave;
+	private Double totalLeave;
+	private Double usedLeave;
 	private Date grantDate;
 	private Date expirationDate;
 	private String userId;
